Create fresh and old bread
A bakery sells loaves of bread for 185 rupees each. Day old bread is discounted by 60 percent. Write a python program that begins by reading the number of loaves of day old bread being purchased from the user. Then your program should display the regular price for the bread, the discount because it is a day old, and the total price. All of the values should be displayed using two decimal places, and the decimal points in all of the numbers should be aligned when reasonable values are entered by the user.
Sample Input:
Enter the number of fresh loves purchased: 5
Enter the number of day old loaves purchased: 3
Sample Output:
Regular price: Rs.185.00
Amount of  new loaves: 925.00
Amount of day old loaves: 333.00
Total amount: Rs. 1258.00