Create Two strings “s” and “t” are isomorphic
Given two strings “s” and “t”, determine if they are isomorphic. Two strings “s” and “t” are isomorphic if the characters in “s” can be replaced to get “t”. All occurrences of a character must be replaced with another character while preserving the order of characters. No two characters may map to the same character, but a character may map to itself.
Constraints
•	s and t consist of any valid ascii character.
Test Cases:
1.Input: s = "egg", t = "add"
      Output: true
2.Input: s = "foo", t = "bar"
      Output: false
3.Input: s = "paper", t = "title"
      Output: true
4.Input: s = "fry", t = "sky"
      Output: true
5. Input: s = "apples", t = "apple"
      Output: false