Create Find two lines, which together with x-axis forms a container
Given n non-negative integers a1,a2,a3,…an where each represents a point at coordinate (i, ai) . ‘ n ‘ vertical lines are drawn such that the two endpoints of line i is at (i, ai) and (i,0). 
Find two lines, which together with x-axis forms a container, such that the container contains the most water. The program should return an integer which corresponds to the maximum area of water that can be contained (maximum area instead of maximum volume sounds weird but this is the 2D plane we are working with for simplicity).
Note: You may not slant the container. 
Test case:
1.Input: array = [1, 5, 4, 3]
Output: 6
2.Input: array = [3, 1, 2, 4, 5]
Output: 12
3.Input: array = [1,8,6,2,5,4,8,3,7]
Output: 49
4.Input: array = [1,1]
Output: 1
5.Input: array = [7,3]
Output: 3