Create the least number of perfect square numbers that sum to n
Given an integer n, return the least number of perfect square numbers that sum to n. A perfect square is an integer that is the square of an integer; in other words, it is the product of some integer with itself. For example, 1, 4, 9, and 16 are perfect squares while 3 and 11 are not.
Test cases:
1.Input: n = 12 output: 3
2. Input: n = 13, Output: 3
3.Input : n= 1, Output: 1
4.Input: n=4, Output: 2
5.Input: n=3, Output: 1