Create checks if reversing the number results in a mirror number
Raju, has again started troubling people in your city. The people have turned on to you for getting rid of Raju. Raju presents to you a number consisting of numbers from 0 to 9 characters. He wants you to reverse it from the final answer such that the number becomes Mirror number. A Mirror is a number which equals its reverse. The hope of people are on you so you have to solve the riddle. You have to tell if some number exists which you would reverse to convert the number into Mirror
Sample input: 
Enter the number: 123456
Sample output:
Mirror image: 654321
Test cases:
1.	Sell123
2.	5489236
3.	Abc-abc
4.	%$$$$^&
5.	-123456