Create displays the shipping charge
An online retailer provides express shipping for many of its items at a rate of 750 for the first item, and 200 for each subsequent item. Write a function that takes the number of items in the order as its only parameter. Return the shipping charge for the order as the functions result. Include the code that reads the number of items purchased from the user and displays the shipping charge. All float values are displayed correct to 2 decimal places. 
Sample Input	
10
Sample Output
2550
Test Cases:
1.	45
2.	-12
3.	12
4.	0
5.	56